pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
name = "simon-fastapi"
version = "0.1.0"
description = ""
authors = ["arlianasrori <arlianasrori@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
fastapi = "^0.111.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.31"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = "^2.8.2"
python-random-strings = "^0.3"
asyncpg = "^0.29.0"
# psycopg2 = "^2.9.8"
bcrypt = "^4.2.0"
alembic = "^1.13.2"
pytz = "^2024.2"
aiofiles = "^24.1.0"
schedule = "^1.2.2"
apscheduler = "^3.10.4"
psycopg2-binary = "^2.9.9"